test(allow-scripts): move and adapt the test case for implicit install with gyp
